% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_fix.R
\name{ba}
\alias{ba}
\title{Functions from gamlss/gamlss.add with bugs fixed}
\usage{
ba(formula, control = ba.control(...), ...)
}
\arguments{
\item{formula}{A formula of the model.}

\item{control}{The control of the model fitting.}

\item{...}{Other arguments.}
}
\description{
An additive function to be used while fitting GAMLSS models. The interface for \code{bam()} in the \pkg{mgcv} package.
}
\section{ba}{
NA
}

