% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_count}
\alias{example_count}
\title{A sparse matrix with example data}
\format{
A dgCMatrix with 100 rows (genes) and 2087 cols (cells)
}
\usage{
data("example_count")
}
\value{
The corresponding sparse matrix object 'example_count'
}
\description{
A sparse matrix with example data
}
\keyword{datasets}
