% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_sce}
\alias{example_sce}
\title{A SingelCellExperiment object containing both cell type and pseudotime}
\format{
A dataset with 10 rows (genes) and 1289 cols (cells)
}
\usage{
data("example_sce")
}
\value{
The corresponding SingleCellExperiment object
}
\description{
A SingelCellExperiment object containing both cell type and pseudotime
}
\keyword{datasets}
