% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_para.R
\name{extract_para}
\alias{extract_para}
\title{Extract the parameters of each cell's distribution}
\usage{
extract_para(
  sce,
  marginal_list,
  n_cores,
  family_use,
  new_covariate,
  parallelization = "mcmapply",
  BPPARAM = NULL,
  data = NULL
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{marginal_list}{A list of fitted regression models from \code{\link{fit_marginal}} for each gene in sce.}

\item{n_cores}{An integer. The number of cores to use.}

\item{family_use}{A string of the marginal distribution.
Must be one of 'poisson', 'nb', 'zip', 'zinb' or 'gaussian', which represent 'poisson distribution',
'negative binomial distribution', 'zero-inflated poisson distribution', 'zero-inflated negative binomail distribution',
and 'gaussian distribution' respectively.}

\item{new_covariate}{A data.frame which contains covaraites of targeted simulated data from  \code{\link{construct_data}} and the
correlation group assignment for each cell in the column 'corr_group'.}

\item{parallelization}{A string indicating the specific parallelization function to use.
Must be one of 'mcmapply', 'bpmapply', or 'pbmcmapply', which corresponds to the parallelization function in the package
\code{parallel},\code{BiocParallel}, and \code{pbmcapply} respectively. The default value is 'mcmapply'.}

\item{BPPARAM}{A \code{MulticoreParam} object or NULL. When the parameter parallelization = 'mcmapply' or 'pbmcmapply',
this parameter must be NULL. When the parameter parallelization = 'bpmapply',  this parameter must be one of the
\code{MulticoreParam} object offered by the package 'BiocParallel. The default value is NULL.}

\item{data}{A dataframe which is used when fitting the gamlss model}
}
\value{
A list with the components:
\describe{
  \item{\code{mean_mat}}{A cell by feature matrix of the mean parameter.}
  \item{\code{sigma_mat}}{A cell by feature matrix of the sigma parameter (for Gaussian, the variance; for NB, the dispersion.).}
  \item{\code{zero_mat}}{A cell by feature matrix of the zero-inflation parameter (only non-zero for ZIP and ZINB).}
}
}
\description{
\code{extract_para} generates parameter matricies which determine each cell's distribution
}
\details{
The function takes the new covariate (if use) from \code{\link{construct_data}} and
marginal models from \code{\link{fit_marginal}}.
}
