% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_marginal.R
\name{fit_marginal}
\alias{fit_marginal}
\title{Fit the marginal models}
\usage{
fit_marginal(
  data,
  predictor = "gene",
  mu_formula,
  sigma_formula,
  family_use,
  n_cores,
  usebam,
  parallelization = "mcmapply",
  BPPARAM = NULL
)
}
\arguments{
\item{data}{An object from \code{\link{construct_data}}. ## Fix later}

\item{predictor}{The predictor for the gam/gamlss model. Default is gene. ## Fix later}

\item{mu_formula}{A string of the mu parameter formula}

\item{sigma_formula}{A string of the sigma parameter formula}

\item{family_use}{A string or a vector of strings of the marginal distribution.
Must be one of 'binomial', 'poisson', 'nb', 'zip', 'zinb' or 'gaussian', which represent 'poisson distribution',
'negative binomial distribution', 'zero-inflated poisson distribution', 'zero-inflated negative binomail distribution',
and 'gaussian distribution' respectively.}

\item{n_cores}{An integer. The number of cores to use.}

\item{usebam}{A logic variable. If use \code{\link[mgcv]{bam}} for acceleration.}

\item{parallelization}{A string indicating the specific parallelization function to use.
Must be one of 'mcmapply', 'bpmapply', or 'pbmcmapply', which corresponds to the parallelization function in the package
\code{parallel},\code{BiocParallel}, and \code{pbmcapply} respectively. The default value is 'mcmapply'.}

\item{BPPARAM}{A \code{MulticoreParam} object or NULL. When the parameter parallelization = 'mcmapply' or 'pbmcmapply',
this parameter must be NULL. When the parameter parallelization = 'bpmapply',  this parameter must be one of the
\code{MulticoreParam} object offered by the package 'BiocParallel. The default value is NULL.}
}
\value{
A list of fitted regression models. The length is equal to the total feature number.
}
\description{
\code{fit_marginal} fits the per-feature regression models.
}
\details{
The function takes the result from \code{\link{construct_data}} as the input,
and fit the regression models for each feature based on users' specification.
}
