% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_fix.R
\name{ga.control}
\alias{ga.control}
\title{Functions from gamlss/gamlss.add with bugs fixed}
\usage{
ga.control(
  offset = NULL,
  method = "REML",
  optimizer = c("outer", "newton"),
  control = list(),
  scale = 0,
  select = FALSE,
  knots = NULL,
  sp = NULL,
  min.sp = NULL,
  H = NULL,
  gamma = 1,
  paraPen = NULL,
  in.out = NULL,
  drop.unused.levels = TRUE,
  drop.intercept = NULL,
  discrete = FALSE,
  ...
)
}
\arguments{
\item{offset}{The offset in the formula.}

\item{method}{The smoothing parameter estimation method.}

\item{optimizer}{An array specifying the numerical optimization method to use to optimize the smoothing parameter estimation criterion (given by \code{method})}

\item{control}{A list of fit control parameters to replace defaults returned by \code{gam.control}.}

\item{scale}{If this is positive then it is taken as the known scale parameter. Negative signals that the scale parameter is unknown. 0 signals that the scale parameter is 1 for Poisson and binomial and unknown otherwise.}

\item{select}{If this is \code{TRUE} then \code{gam()} can add an extra penalty to each term so that it can be penalized to zero.}

\item{knots}{This is an optional list containing user specified knot values to be used for basis construction.}

\item{sp}{A vector of smoothing parameters can be provided here.}

\item{min.sp}{Lower bounds can be supplied for the smoothing parameters.}

\item{H}{A user supplied fixed quadratic penalty on the parameters of the GAM can be supplied, with this as its coefficient matrix.}

\item{gamma}{Increase this beyond 1 to produce smoother models.}

\item{paraPen}{Optional list specifying any penalties to be applied to parametric model terms.}

\item{in.out}{Optional list for initializing outer iteration.}

\item{drop.unused.levels}{By default unused levels are dropped from factors before fitting. For some smooths involving factor variables you might want to turn this off.}

\item{drop.intercept}{Set to \code{TRUE} to force the model to really not have the a constant in the parametric model part, even with factor variables present. Can be vector when \code{formula} is a list.}

\item{discrete}{Experimental option for setting up models for use with discrete methods employed in \code{bam}.}

\item{...}{Other arguments}
}
\value{
A control object
}
\description{
The control for \code{ga()}. From \code{gamlss.add::ga.control()} and \code{gamlss::gam()}.
}
\section{ga.control}{
NA
}

