% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_fix.R
\name{ga}
\alias{ga}
\title{Functions from gamlss/gamlss.add with bugs fixed}
\usage{
ga(formula, envir, control = ga.control(...), ...)
}
\arguments{
\item{formula}{A formula of the model.}

\item{envir}{The environment.}

\item{control}{The control of the model fitting.}

\item{...}{Other arguments.}
}
\description{
Functions from gamlss/gamlss.add with bugs fixed
}
\section{ga}{
NA
}

