% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_fix.R
\name{model.frame.gamlss}
\alias{model.frame.gamlss}
\title{Functions from gamlss/gamlss.add with bugs fixed}
\usage{
\method{model.frame}{gamlss}(
  formula,
  what = c("mu", "sigma", "nu", "tau"),
  parameter = NULL,
  ...
)
}
\arguments{
\item{formula}{A formula of the model.}

\item{what}{For which parameter to extract the model.frame, terms or model.frame.}

\item{parameter}{Equivalent to \code{what}.}

\item{...}{Other arguments.}
}
\value{
a vector or matrix of predicted values.
}
\description{
The gamlss versions of the generic function \code{model.frame}
}
\section{model.frame.gamlss}{
NA
}

