% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_fix.R
\name{predict.gamlss}
\alias{predict.gamlss}
\title{fix bugs for the original predict.gamlss}
\usage{
\method{predict}{gamlss}(
  object,
  what = c("mu", "sigma", "nu", "tau"),
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  terms = NULL,
  se.fit = FALSE,
  data = NULL,
  ...
)
}
\arguments{
\item{object}{A gamlss object}

\item{what}{to be added}

\item{parameter}{to be added}

\item{newdata}{data frame}

\item{type}{to be added}

\item{terms}{to be added}

\item{se.fit}{to be added}

\item{data}{to be added}
}
\value{
A num.
}
\description{
fix bugs for the original predict.gamlss
}
