% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pseudotime}
\alias{pseudotime}
\title{A numeric vector of pseudotime values}
\format{
A numeric vector with 2087 values
}
\usage{
data("pseudotime")
}
\value{
The corresponding pseudotime of the 'example_count'
}
\description{
A numeric vector of pseudotime values
}
\keyword{datasets}
