% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_new.R
\name{simu_new}
\alias{simu_new}
\title{Simulate new data}
\usage{
simu_new(
  sce,
  mean_mat,
  sigma_mat,
  zero_mat,
  quantile_mat,
  family_use,
  n_cores,
  new_covariate
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{mean_mat}{A cell by feature matrix of the mean parameter.}

\item{sigma_mat}{A cell by feature matrix of the sigma parameter.}

\item{zero_mat}{A cell by feature matrix of the zero-inflation parameter.}

\item{quantile_mat}{A cell by feature matrix of the multivariate quantile.}

\item{family_use}{A string of the marginal distribution.
Must be one of 'poisson', 'nb', 'zip', 'zinb' or 'gaussian'.}

\item{n_cores}{An integer. The number of cores to use.}

\item{new_covariate}{A data.frame which contains covariates of targeted simulated data from  \code{\link{construct_data}}.}
}
\value{
A feature by cell matrix of the new simulated count (expression) matrix.
}
\description{
\code{simu_new} generates new simulated data based on fitted marginal and copula models.
}
\details{
The function takes the new covariate (if use) from \code{\link{construct_data}},
parameter matricies from \code{\link{extract_para}} and multivariate Unifs from \code{\link{fit_copula}}.
}
