% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_cov.R
\name{sparse_cov}
\alias{sparse_cov}
\title{This function computes the thresholding sparse covariance/correlation estimator 
with the optimal threshold level.}
\usage{
sparse_cov(
  data,
  method = c("cv", "qiu"),
  operator = c("hard", "soft", "scad", "al"),
  corr = TRUE
)
}
\arguments{
\item{data}{The data matrix.}

\item{method}{The choice of method to select the optimal threshold level.}

\item{operator}{The choice of the thresholding operator.}

\item{corr}{The indicator of computing correlation or covariance matrix.}
}
\value{
The thresholding sparse covariance/correlation estimator.
}
\description{
Part from Chenxin Jiang
}
\examples{
print("No example")
}
