% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_copula.R
\name{fit_copula}
\alias{fit_copula}
\title{Fit the copula model}
\usage{
fit_copula(
  sce,
  assay_use,
  input_data,
  new_covariate = NULL,
  marginal_list,
  family,
  copula = "vine",
  DT = TRUE,
  pseudo_obs = FALSE,
  epsilon = 1e-06,
  family_set = c("gaussian", "indep"),
  n_cores
)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{assay_use}{A string which indicates the assay you will use in the sce.
Default is 'counts'.}

\item{input_data}{A input count matrix.}

\item{new_covariate}{A data.frame which contains covaraites of targeted simulated data from  \code{\link{construct_data}}.}

\item{marginal_list}{A list of fitted regression models from \code{\link{fit_marginal}}.}

\item{family}{A string of the marginal distribution. Must be one of 'poisson', 'nb', 'zip', 'zinb' or 'gaussian'.}

\item{copula}{A string of the copula choice. Must be one of 'gaussian' or 'vine'. Default is 'vine'.}

\item{DT}{A logic variable. If TRUE, perform the distributional transformation
to make the discrete data 'continuous'. This is useful for discrete distributions (e.g., Poisson, NB).
Default is TRUE.}

\item{pseudo_obs}{A logic variable. If TRUE, use the empirical quantiles instead of theoretical quantiles for fitting copula.
Default is FALSE.}

\item{epsilon}{A numeric variable for preventing the transformed quantiles to collapse to 0 or 1.}

\item{family_set}{A string or a string vector of the bivarate copula families. Default is c("gaussian", "indep").}

\item{n_cores}{An integer. The number of cores to use.}
}
\value{
A list with the components:
\describe{
  \item{\code{new_mvu}}{A matrix of the new multivariate uniform distribution from the copula.}
  \item{\code{corr_list}}{A list of the fitted copula model. If using Gaussian copula, a list of correlation matrices; if vine, a list of vine objects.}
  \item{\code{model_aic}}{A vector of the marginal AIC and the copula AIC.}
  \item{\code{model_bic}}{A vector of the marginal BIC and the copula BIC.}
}
}
\description{
\code{fit_copula} fits the copula model.
}
\details{
This function takes the result from \code{\link{fit_marginal}} as the input and
and fit the copula model on the residuals.
}
