% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_fix.R
\name{gamlss.ba}
\alias{gamlss.ba}
\title{Support for Function ba()}
\usage{
gamlss.ba(x, y, w, xeval = NULL, ...)
}
\arguments{
\item{x}{The explanatory variables}

\item{y}{Iterative y variable}

\item{w}{Iterative weights}

\item{xeval}{If xeval=TRUE then predicion is used}

\item{...}{Other arguments}
}
\description{
This is support for the  smoother functions \code{ba()} intefaces for Simon Woood's \code{bam()} functions from package \pkg{mgcv}. It is not intended to be called directly by users. From \code{gamlss.add::gamlss.ba}.
}
