% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_lrt.R
\name{perform_lrt}
\alias{perform_lrt}
\title{Perform the likelihood ratio test}
\usage{
perform_lrt(alter_marginal, null_marginal)
}
\arguments{
\item{alter_marginal}{A list of marginal models from the alternative hypothesis.}

\item{null_marginal}{A list of marginal models from the null hypothesis. It must be strictly nested in the alternative model.}
}
\value{
A data.frame of the LRT result.
}
\description{
\code{perform_lrt} performs the likelihood ratio test to compare two marginal models.
}
\details{
The function takes two lists of marginal models (by default, the first list is the alternative and the second is the null)
from \code{\link{fit_marginal}}.
}
