% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlss_fix.R
\name{predict.gamlss}
\alias{predict.gamlss}
\title{Functions from gamlss/gamlss.add with bugs fixed}
\usage{
\method{predict}{gamlss}(
  object,
  what = c("mu", "sigma", "nu", "tau"),
  parameter = NULL,
  newdata = NULL,
  type = c("link", "response", "terms"),
  terms = NULL,
  se.fit = FALSE,
  data = NULL,
  ...
)
}
\arguments{
\item{object}{A model fitted using GAMLSS.}

\item{what}{which distribution parameter is required, the default is "mu".}

\item{parameter}{equivalent to the parameter \code{what}.}

\item{newdata}{a dataframe containing all covariates used in the fitted model.}

\item{type}{the default, gets the linear predictor for the specified distribution parameter.}

\item{terms}{if \code{type="terms"}, which terms to be selected (default is all terms).}

\item{se.fit}{if TRUE the approximate standard errors of the appropriate type are extracted if exist.}

\item{data}{the dataframe used in the original fit if is not defined in the call.}

\item{...}{Other arguments.}
}
\value{
A vector or matrix of predicted values.
}
\description{
The GAMLSS specific method which produce predictors for a new data set for a specified parameter from a GAMLSS objects.
}
\section{predict.gamlss}{
NA
}

