% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu_new.R
\name{simu_new}
\alias{simu_new}
\title{Simulate new data}
\usage{
simu_new(sce, marginal_list, copula_list, n_cores, family, new_covariate)
}
\arguments{
\item{sce}{A \code{SingleCellExperiment} object.}

\item{marginal_list}{A list of fitted regression models from \code{\link{fit_marginal}}.}

\item{copula_list}{A list of fitted copula models from \code{\link{fit_copula}}.}

\item{n_cores}{An integer. The number of cores to use.}

\item{family}{A string of the marginal distribution.
Must be one of 'poisson', 'nb', 'zip', 'zinb' or 'gaussian'.}

\item{new_covariate}{A data.frame which contains covaraites of targeted simulated data from  \code{\link{construct_data}}.}
}
\value{
A list with the components:
\describe{
  \item{\code{new_count}}{A matrix of the new simulated count (expression) matrix.}
  \item{\code{cor_list}}{A list of the fitted copula model. If using Gaussian copula, a list of correlation matrices; if vine, a list of vine objects.}
  \item{\code{model_aic}}{A vector of the marginal AIC and the copula AIC.}
}
}
\description{
\code{simu_new} generates new simulated data based on fitted marginal and copula models.
}
\details{
The function takes the new covariate (if use) from \code{\link{construct_data}},
marginal models from \code{\link{fit_marginal}} and multivariate Unifs from \code{\link{fit_copula}}.
}
